refactor(src/modules): refactor url module to abstratic repository
