+++ src/shared/container/index.ts
@@ -1,7 +1,7 @@
 import { container } from 'tsyringe';
 
-import { UrlRepository } from '../infra/database/repositories/implementations/UrlRepository';
-import { IUrlRepository } from '../infra/database/repositories/IUrlRepository';
+import { UrlRepository } from '../../modules/url/infra/typeorm/repositories/UrlRepository';
+import { IUrlRepository } from '../../modules/url/repositories/IUrlRepository';
 
 container.registerSingleton<IUrlRepository>('UrlRepository', UrlRepository);
